+++ src/main/java/com/tallyto/gestorfinanceiro/config/database/JpaConfig.java
@@ -1,0 +1,10 @@
+package com.tallyto.gestorfinanceiro.config.database;
+
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing(auditorAwareRef = "auditorProvider")
+public class JpaConfig {
+}