+++ src/main/java/com/tallyto/gestorfinanceiro/config/security/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.tallyto.gestorfinanceiro.config;
+package com.tallyto.gestorfinanceiro.config.security;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
@@ -8,6 +8,9 @@ import org.springframework.security.config.annotation.authentication.configurati
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+import com.tallyto.gestorfinanceiro.config.JwtAuthenticationFilter;
+
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -26,6 +29,11 @@ public class SecurityConfig {
             .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
             .authorizeHttpRequests(auth -> auth
                 .requestMatchers("api/auth/**").permitAll()
+                .requestMatchers("api/tenants/cadastro").permitAll()
+                .requestMatchers("api/tenants/verificar").permitAll()
+                .requestMatchers("api/tenants/confirmar").permitAll()
+                .requestMatchers("api/usuarios").permitAll()
+                .requestMatchers("api/tenants/verificar-dominio").permitAll()
                 .anyRequest().authenticated()
             )
             .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class);